Add the ai_types crate and move pure AI model types into it
Create a new crates/ai_types crate that holds pure AI model types with a
small dependency budget (serde, uuid, anyhow, thiserror). Move
AIConversationId, AIAgentActionId, TaskId, AmbientAgentTaskId,
EntrypointType, PassiveSuggestionTriggerType, and WarpAiExecutionContext
into it. Move SkillDescriptor into the ai crate, because it uses ai and
warp_core types.

The warp crate re-exports the moved types from their old paths, so most
call sites do not change. The persistence conversions move into the
persistence crate to satisfy the orphan rule, and
WarpAiExecutionContext::new becomes the free function
execution_context_for_session because it reads app Session state.
