+++ app/src/ai/ambient_agents/mod.rs
@@ -1,9 +1,3 @@
-use std::fmt::Display;
-use std::str::FromStr;
-
-use serde::{Deserialize, Serialize};
-use uuid::{NonNilUuid, Uuid};
-
 use crate::ai::agent::conversation::{AIConversation, ConversationStatus};
 use crate::ai::agent::{
     AIAgentOutputStatus, CancellationReason, FinishedAIAgentOutput, RenderableAIError,
@@ -26,34 +20,7 @@ pub const OUT_OF_CREDITS_TASK_FAILURE_MESSAGE: &str =
 pub const SERVER_OVERLOADED_TASK_FAILURE_MESSAGE: &str =
     "Warp is temporarily overloaded. Please try again shortly.";
 
-#[derive(Debug, thiserror::Error)]
-#[error("Invalid task ID: {0}")]
-pub struct ParseAmbientAgentTaskIdError(#[from] uuid::Error);
-
-/// A globally unique ID for an ambient agent task.
-#[derive(Serialize, Deserialize, Debug, Clone, Copy, PartialEq, Eq, Hash)]
-pub struct AmbientAgentTaskId(NonNilUuid);
-
-impl Display for AmbientAgentTaskId {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        self.0.fmt(f)
-    }
-}
-
-impl FromStr for AmbientAgentTaskId {
-    type Err = ParseAmbientAgentTaskIdError;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        let uuid = Uuid::try_parse(s)?;
-        Ok(Self(NonNilUuid::try_from(uuid)?))
-    }
-}
-
-impl From<AmbientAgentTaskId> for cynic::Id {
-    fn from(id: AmbientAgentTaskId) -> Self {
-        Self::new(id.to_string())
-    }
-}
+pub use ai_types::AmbientAgentTaskId;
 
 /// High-level outcome of an ambient agent conversation.
 #[derive(Clone, Debug)]
