+++ app/src/ai/skills/mod.rs
@@ -53,11 +53,9 @@ impl ActiveSkillLookupError {
 
 #[cfg(not(target_family = "wasm"))]
 mod global_skills;
+pub use ai::skills::SkillDescriptor;
 #[cfg(not(target_family = "wasm"))]
 pub use global_skills::{filter_skills_by_spec, resolve_skill_repos};
-
-mod listed_skill;
-pub use listed_skill::SkillDescriptor;
 
 mod skill_utils;
 pub use skill_utils::{
