+++ app/src/terminal/model/session/active_session.rs
@@ -8,7 +8,9 @@ use warp_util::standardized_path::StandardizedPath;
 use warpui::{AppContext, Entity, ModelContext, ModelHandle};
 
 use super::{Session, SessionType, Sessions};
-use crate::ai_assistant::execution_context::WarpAiExecutionContext;
+use crate::ai_assistant::execution_context::{
+    WarpAiExecutionContext, execution_context_for_session,
+};
 use crate::terminal::ShellLaunchData;
 use crate::terminal::model::session::SessionsEvent;
 use crate::terminal::model_events::{ModelEvent, ModelEventDispatcher};
@@ -126,7 +128,9 @@ impl ActiveSession {
 
     /// Returns the `WarpAiExecutionContext` for the active session.
     pub fn ai_execution_environment(&self, app: &AppContext) -> Option<WarpAiExecutionContext> {
-        self.session(app).as_ref().map(WarpAiExecutionContext::new)
+        self.session(app)
+            .as_ref()
+            .map(execution_context_for_session)
     }
 }
