+++ crates/ai/src/skills/listed_skill.rs
@@ -1,6 +1,7 @@
-use ai::skills::{ParsedSkill, SkillProvider, SkillReference, SkillScope};
 use serde::{Deserialize, Serialize};
 use warp_core::ui::icons::Icon;
+
+use super::{ParsedSkill, SkillProvider, SkillReference, SkillScope};
 
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, Hash)]
 pub struct SkillDescriptor {
