+++ crates/ai/src/skills/mod.rs
@@ -1,4 +1,5 @@
 mod conversion;
+mod listed_skill;
 mod parse_skill;
 mod parser;
 mod read_skills;
@@ -8,6 +9,7 @@ pub use conversion::{
     SkillConversionError, SkillPathOrigin, skill_reference_from_api_skill_ref,
     skill_reference_from_read_skill_ref,
 };
+pub use listed_skill::SkillDescriptor;
 pub use parse_skill::{
     ParsedSkill, parse_bundled_skill, parse_skill, parse_skill_content_at_location,
 };
