+++ crates/ai_types/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "ai_types"
+version = "0.1.0"
+edition = "2024"
+publish.workspace = true
+license.workspace = true
+
+[dependencies]
+anyhow.workspace = true
+serde.workspace = true
+serde_json.workspace = true
+thiserror.workspace = true
+uuid.workspace = true
