+++ crates/ai_types/src/ambient_agent.rs
@@ -1,0 +1,28 @@
+use std::fmt::Display;
+use std::str::FromStr;
+
+use serde::{Deserialize, Serialize};
+use uuid::{NonNilUuid, Uuid};
+
+#[derive(Debug, thiserror::Error)]
+#[error("Invalid task ID: {0}")]
+pub struct ParseAmbientAgentTaskIdError(#[from] uuid::Error);
+
+/// A globally unique ID for an ambient agent task.
+#[derive(Serialize, Deserialize, Debug, Clone, Copy, PartialEq, Eq, Hash)]
+pub struct AmbientAgentTaskId(NonNilUuid);
+
+impl Display for AmbientAgentTaskId {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        self.0.fmt(f)
+    }
+}
+
+impl FromStr for AmbientAgentTaskId {
+    type Err = ParseAmbientAgentTaskIdError;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        let uuid = Uuid::try_parse(s)?;
+        Ok(Self(NonNilUuid::try_from(uuid)?))
+    }
+}
